% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hijazi2020.R
\docType{data}
\name{hijazi2020}
\alias{hijazi2020}
\title{List of human kinase-substrate relationships measured by Hijazi et al., 2020.}
\format{
A \code{data.table} of 19287 K-S interactions.
}
\source{
\url{https://doi.org/10.1038/s41587-019-0391-9}
}
\usage{
hijazi2020
}
\description{
This data set contains a list of human kinase-substrate relationships measured by Hijazi et al., 2020.
}
\examples{
\dontrun{
data(hijazi2020)
summary(hijazi2020)
}
}
\references{
Hijazi, M., Smith, R., Rajeeve, V. et al. Reconstructing kinase network topologies from phosphoproteomics data reveals cancer-associated rewiring. Nat Biotechnol 38, 493–502 (2020). https://doi.org/10.1038/s41587-019-0391-9
}
\author{
George Rosenberger
}
\keyword{datasets}
