% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmd.R
\docType{data}
\name{hsmd}
\alias{hsmd}
\title{List of human kinase/phosphatase-substrate interactions predicted by HSM/D}
\format{
A \code{data.table} of 14165303 K/P-S interactions.
}
\source{
\url{https://figshare.com/articles/Predictions_-_Domain-Peptide_and_Protein-Protein_Interactions_-_HSM/10084745/2}
}
\usage{
hsmd
}
\description{
This data set contains a list of human kinase/phosphatase-substrate interactions based on the domain-level HSM/D predictions.
}
\examples{
\dontrun{
data(hsmd)
summary(hsmd)
}
}
\references{
Cunningham JM, Koytiger G, Sorger PK, AlQuraishi M. Nat Methods. 2020 Feb;17(2):175-183. doi: 10.1038/s41592-019-0687-1.
}
\author{
George Rosenberger
}
\keyword{datasets}
