% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/johnson2023.R
\docType{data}
\name{johnson2023}
\alias{johnson2023}
\title{List of human kinase-substrate relationships measured by Johnson et al., 2023.}
\format{
A \code{data.table} of 26764293 K-S interactions.
}
\source{
\url{https://doi.org/10.1038/s41586-022-05575-3}
}
\usage{
johnson2023
}
\description{
This data set contains a list of human kinase-substrate relationships measured by Johnson et al., 2023.
}
\examples{
\dontrun{
data(johnson2023)
summary(johnson2023)
}
}
\references{
Johnson, J.L., Yaron, T.M., Huntsman, E.M. et al. An atlas of substrate specificities for the human serine/threonine kinome. Nature 613, 759–766 (2023).
}
\author{
George Rosenberger
}
\keyword{datasets}
