% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinases.R
\docType{data}
\name{kinases}
\alias{kinases}
\title{List of human kinases (UniProtKB/Swiss-Prot)}
\format{
A \code{vector} of 504 UniProtKB/Swiss-Prot identifiers.
}
\source{
\url{https://www.uniprot.org/docs/pkinfam}
}
\usage{
kinases
}
\description{
This data set contains a list of human kinases based on pkinfam.txt version 2018_03.
}
\examples{
\dontrun{
data(kinases)
summary(kinases)
}
}
\author{
George Rosenberger
}
\keyword{datasets}
