% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsdb.R
\docType{data}
\name{lpsdb}
\alias{lpsdb}
\title{List of human kinase/phosphatase-substrate interactions predicted by LinkPhinder (substrate-level).}
\format{
A \code{data.table} of 11581940 K/P-S interactions.
}
\source{
\url{https://doi.org/10.1371/journal.pcbi.1007578}
}
\usage{
lpsdb
}
\description{
This data set contains a list of human kinase/phosphatase-substrate interactions based on the substrate-level LinkPhinder predictions.
}
\examples{
\dontrun{
data(lpsdb)
summary(lpsdb)
}
}
\references{
Nováček V, McGauran G, Matallanas D, Vallejo Blanco A, Conca P, Muñoz E, Costabello L, Kanakaraj K, Nawaz Z, Walsh B, Mohamed SK, Vandenbussche PY, Ryan CJ, Kolch W, Fey D. PLoS Comput Biol. 2020 Dec 3;16(12):e1007578. doi: 10.1371/journal.pcbi.1007578. eCollection 2020 Dec.
}
\author{
George Rosenberger
}
\keyword{datasets}
