% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdb.R
\docType{data}
\name{pcdb}
\alias{pcdb}
\title{List of human kinase/phosphatase-substrate interactions collected by Pathway Commons, version 12.}
\format{
A \code{data.table} of 17113 K/P-S interactions.
}
\source{
\url{https://www.pathwaycommons.org}
}
\usage{
pcdb
}
\description{
This data set contains a list of human kinase/phosphatase-substrate interactions based on the protein-level Pathway Commons database.
}
\examples{
\dontrun{
data(pcdb)
summary(pcdb)
}
}
\references{
Rodchenkov I, Babur O, Luna A, Aksoy BA, Wong JV, Fong D, Franz M, Siper MC, Cheung M, Wrana M, Mistry H, Mosier L, Dlin J, Wen Q, O'Callaghan C, Li W, Elder G, Smith PT, Dallago C, Cerami E, Gross B, Dogrusoz U, Demir E, Bader GD, Sander C. Nucleic Acids Res. 2020 Jan 8;48(D1):D489-D497. doi: 10.1093/nar/gkz946.
}
\author{
George Rosenberger
}
\keyword{datasets}
