% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphatases.R
\docType{data}
\name{phosphatases}
\alias{phosphatases}
\title{List of human phosphatases (UniProtKB/Swiss-Prot)}
\format{
A \code{vector} of 223 UniProtKB/Swiss-Prot identifiers.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/23674824}
}
\usage{
phosphatases
}
\description{
This data set contains a list of human phosphatases based on the reference below.
}
\examples{
\dontrun{
data(phosphatases)
summary(phosphatases)
}
}
\references{
Li X, Wilmanns M, Thornton J, Köhn M. Sci Signal. 2013 May 14;6(275):rs10. doi: 10.1126/scisignal.2003203.
}
\author{
George Rosenberger
}
\keyword{datasets}
