% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphoviper.db.R
\docType{package}
\name{phosphoviper.db}
\alias{phosphoviper.db}
\title{phosphoviper.db companion data package for phosphoviper}
\description{
This package contains necessary reference databases for phosphoviper.
}
\details{
\tabular{ll}{
Package: \tab phosphoviper.db\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
License: \tab Columbia University Non-Commercial License\cr
}
}
\examples{
# Please see the datasets
}
\author{
George Rosenberger \email{gr2578@cumc.columbia.edu}
}
