% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vespa.db.R
\docType{package}
\name{vespa.db}
\alias{vespa.db}
\title{vespa.db companion data package for vespa}
\description{
This package contains necessary reference databases for vespa.
}
\details{
\tabular{ll}{
Package: \tab vespa.db\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
License: \tab Columbia University Non-Commercial License\cr
}
}
\examples{
# Please see the datasets
}
\author{
George Rosenberger \email{gr2578@cumc.columbia.edu}
}
